HR Analytics Random Forest - RandomSearchCV
A company which is active in Big Data and Data Science wants to hire data scientists among people who successfully pass some courses which conduct by the company. Many people signup for their training. Company wants to know which of these candidates are really wants to work for the company after training or looking for a new employment because it helps to reduce the cost and time as well as the quality of training or planning the courses and categorization of candidates. Information related to demographics, education, experience are in hands from candidates signup and enrollment.
This dataset designed to understand the factors that lead a person to leave current job for HR researches too. By model(s) that uses the current credentials,demographics,experience data you will predict the probability of a candidate to look for a new job or will work for the company, as well as interpreting affected factors on employee decision.
The whole data divided to train and test . Target isn't included in test but the test target values data file is in hands for related tasks. A sample submission correspond to enrollee_id of test set provided too with columns : enrollee _id , target
Note:
•	The dataset is imbalanced.
•	Most features are categorical (Nominal, Ordinal, Binary), some with high cardinality.
•	Missing imputation can be a part of your pipeline as well.
Features
•	enrollee_id : Unique ID for candidate
•	city: City code
•	city_ development _index : Developement index of the city (scaled)
•	gender: Gender of candidate
•	relevent_experience: Relevant experience of candidate
•	enrolled_university: Type of University course enrolled if any
•	education_level: Education level of candidate
•	major_discipline :Education major discipline of candidate
•	experience: Candidate total experience in years
•	company_size: No of employees in current employer's company
•	company_type : Type of current employer
•	lastnewjob: Difference in years between previous job and current job
•	training_hours: training hours completed
•	target: 0 – Not looking for job change, 1 – Looking for a job change
Inspiration
•	Predict the probability of a candidate will work for the company
•	Interpret model(s) such a way that illustrate which features affect candidate decision